Change the Id primitive type to long

+++ RestWithASPNET5_Person/RestWithASPNET5/Model/Book.cs
@@ -2,10 +2,10 @@
 {
     public class Book
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Author { get; set; }
         public DateTime Date { get; set; }
-        public decimal Price { get; set; }
+        public double Price { get; set; }
         public string Title { get; set; }
     }
 }
